test(nika-cadence): the mirror and planner pins get their own file
The W1 tests pushed src/tests.rs to 1,608 lines — over the 1,500 file
cap. The module becomes tests/mod.rs + tests/planner.rs (the
prev_before mirror and the due planner pins, the inverse-law proptest
included); the helpers (at · zoned · utc) stay in the parent, the
children ride super::. The trap-③ static guard's file list follows the
move and gains due.rs.
